Do not include indicator metadata in getStaticProps
Bundle indicators.js into the code. We already have this bundled for use 
in the navbar anyway, so it does not make sense to also include it in 
the page data. In the future, as indicators.js grows in size, it will 
make sense to split the file into the essential metadata (label & slug) 
and the details (bounds, legend, ...) that we only include when 
rendering the map.